+++ gitadora@asphyxia/handlers/MusicList.ts
@@ -1,8 +1,6 @@
 import { getVersion } from "../utils";
-import { processData as ExchainMusic } from "../data/Exchain"
-import { processData as MatixxMusic } from "../data/Matixx"
-import { processData as NextageMusic } from "../data/Nextage"
-import { CommonMusicDataField, readJSONOrXML, readXML } from "../data/helper";
+import { processDataBuilder } from "../data"
+import { CommonMusicDataField, readJSONOrXML, readXML } from "../data";
 
 export const playableMusic: EPR = async (info, data, send) => {
   const version = getVersion(info);
@@ -61,13 +59,7 @@ export const playableMusic: EPR = async (info, data, send) => {
   }
 
   if (music.length == 0) {
-    if (version == 'nextage') {
-      music = _.get(await NextageMusic(), 'music', []);
-    } else if (version == 'exchain') {
-      music = _.get(await ExchainMusic(), 'music', []);
-    } else {
-      music = _.get(await MatixxMusic(), 'music', []);
-    }
+      music = _.get(await processDataBuilder(version), 'music', []);
   }
